# Conflicts:
#   ros2/src/hardware/interface/march_hardware/CMakeLists.txt
#   ros2/src/hardware/interface/march_hardware/test/pressure_sole/pressure_sole_test.cpp

+++ ros2/src/hardware/interface/march_hardware/test/mocks/mock_encoder.h
@@ -20,7 +20,7 @@ public:
 
     MOCK_CONST_METHOD1(positionIUToRadians, double(double));
     MOCK_CONST_METHOD1(velocityIUToRadians, double(double));
-    MOCK_CONST_METHOD1(positionRadiansToIU, double(double));
+    MOCK_CONST_METHOD1(positionRadiansToIU, int32_t(double));
     MOCK_CONST_METHOD1(velocityRadiansToIU, double(double));
 };
 // NOLINTEND
